:truck: check token api 폴더 이동 #68
