+++ src/features/auth/index.ts
@@ -1,1 +1,2 @@
 export * from './hooks';
+export { checkToken } from './api';
